submission game

+++ app/routers/topic.py
@@ -1,22 +1,33 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException
 
-from app.schemas.topic import *
-from app.dependencies import db_dep
-from app.models import *
+from app.dependencies import db_dep, current_user_dep, admin_user_dep, staff_user_dep
+from app.models import Topic
+from app.schemas import TopicCreate, TopicResponse
 
 
-router = APIRouter()
+router = APIRouter(prefix="/topics", tags=["topics"])
 
-@router.get('/get', response_model=List[TopicGet])
-def get_topics(db: db_dep):
+
+
+@router.get("/", response_model=list[TopicResponse])
+async def get_topics(db: db_dep):
     return db.query(Topic).all()
 
 
-@router.post('/create', response_model=TopicCreateResponse)
-def create_topic(db: db_dep,
-                 topic: TopicCreate):
-
-    db_topic = Topic(name=topic.name)
+@router.post("/create/", response_model=TopicResponse)
+async def create_topic(
+    topic: TopicCreate,
+    db: db_dep,
+    admin_user: admin_user_dep
+):
+    if db.query(Topic).filter(Topic.name == topic.name).first():
+        raise HTTPException(
+            status_code=400,
+            detail="Topic with this name already exists."
+        )
+    db_topic = Topic(
+        name=topic.name
+    )
 
     db.add(db_topic)
     db.commit()
@@ -25,11 +36,20 @@ def create_topic(db: db_dep,
     return db_topic
 
 
-@router.delete('/delete/{id}', response_model=TopicGet)
-def delete_topic(db: db_dep, id: int):
+@router.delete("/delete/{id}")
+async def delete_topic(id: int, db: db_dep, admin_user: admin_user_dep):
     db_topic = db.query(Topic).filter(Topic.id == id).first()
+
+    if not db_topic:
+        raise HTTPException(
+            status_code=404,
+            detail="Topic not found."
+        )
 
     db.delete(db_topic)
     db.commit()
 
-    return db_topic
+    return {
+        "topic_id": id,
+        "message": "Topic deleted."
+    }
